Added files via upload
You would develop a target application which runs queries on the Yelp data and extracts useful information.The primary users for this application will be potential customers seeking for businesses that match their search criteria. Your application will have a user interface that provides the user the available business categories (main and sub-categories) and the attributes associated with each business category. Using this application the user will search for the businesses from various business categories that have the properties (attributes) the user is looking for. The user can filter the search results by state, city or zipcode. In addition, the user may choose to view only the businesses that are open or closed on certain days of the week. The application should also allow the user to view the reviews provided for each business.
You will be designing your application a standalone Java application.

 Populate your database with the Yelp data. Generate INSERT statements for your